backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
config==0.5.1
Django==4.0.4
django-ninja==0.17.0
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
idna==3.3
jose==1.0.0
numpy==1.23.0
pandas==1.4.3
psycopg2==2.9.3
passlib==1.7.4
pyasn1==0.4.8
pycosat==0.6.3
pydantic==1.9.1
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2022.1
requests==2.28.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
